reorganizing my applications into different compartments so not all the logic is just in one file, helps find errors by doing this
